Move helper method to TransientStorageStrategy
Also move etag generation since we cannot reuse Payloads in the
general case.
